🐖🙎🏼‍♀️ ↝ pretty happy with this refactoring. [ SGV2-8 ]
