+++ components/Content/[id]/BasePlanetSector.tsx
@@ -1,9 +1,9 @@
 import { useSession, useSupabaseClient } from "@supabase/auth-helpers-react";
 import { useRouter } from "next/router";
 import React, { useEffect, useState } from "react";
-import RoverImageGallery, { RoverImage, RoverImageNoHandle } from "../PlanetData/RandomRoverImage";
-import StructureComponent, { PlacedStructures } from "../Sectors/StructureCreate";
-import { SectorStructureOwned } from "../../Inventory/UserOwnedItems";
+import RoverImageGallery, { RoverImage, RoverImageNoHandle } from "../Planets/PlanetData/RandomRoverImage";
+import StructureComponent, { PlacedStructures } from "../Planets/Sectors/StructureCreate";
+import { SectorStructureOwned } from "../Inventory/UserOwnedItems";
 // import { CreateMenuBar, SectorCircularMenu } from "../../../Core/BottomBar";
 
 const AddResourceToInventory = ({ resource, sectorId }) => {
