+++ components/Content/[id]/IndividualBasePlanet.tsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useRouter } from "next/router";
 import { useSupabaseClient, useSession } from "@supabase/auth-helpers-react";
 import Link from "next/link";
-import { LightkurveBaseGraph } from "../PlanetData/ContentPlaceholder";
+import { LightkurveBaseGraph } from "../Planets/PlanetData/ContentPlaceholder";
 
 export function IndividualBasePlanetDesktop({ id }: { id: string }) { 
   const router = useRouter();
