+++ components/Gameplay/Explore/RandomImage.tsx
