+++ components/Gameplay/Explore/RoverExplore.tsx
