+++ components/Gameplay/Sandbox/SectorStructures.tsx
